io/requirements.txt
opensearch-py
sentence-transformers
pyyaml
python-telegram-bot
transformers>=4,<5
torch
accelerate>=0.27